Extract file assertion helpers in their own module
